To - Do- List
#BOOTSTRAP

Add Item: This feature allows you to add new items to your to-do list. Typically, you would have an input field where you can enter the task description, and a button to submit the new item. Once submitted, the item gets added to the list.

Delete Button: Each item in the to-do list has a delete button associated with it. When clicked, the delete button removes the corresponding item from the list. This feature allows you to remove completed or unwanted tasks.

Search: The search functionality enables you to find specific items in your to-do list. You can have an input field where you can enter a keyword or a part of the task description. As you type, the list dynamically updates to show only the items that match the search criteria.